fix copyfile.py which is used on windows, because it may cause an error when the directory is denied to remove,test=develop

+++ cmake/copyfile.py
@@ -26,18 +26,25 @@ def main():
         dst = os.path.join(dst, pathList[-1])
         if not os.path.exists(dst):
             shutil.copytree(src, dst)
-            print("first copy directory: {0} --->>> {1}".format(src, dst))
+            print("copying directory: {0} -> {1}".format(src, dst))
         else:
-            shutil.rmtree(dst)
-            shutil.copytree(src, dst)
-            print("overwritten copy directory: {0} --->>> {1}".format(src, dst))
+            print(
+                "abandon copying! because directory {0} has already been copied to {1}".
+                format(src, dst))
     else:  #copy file, wildcard
         if not os.path.exists(dst):
             os.makedirs(dst)
         srcFiles = glob.glob(src)
         for srcFile in srcFiles:
-            shutil.copy(srcFile, dst)
-            print("copy file: {0} --->>> {1}".format(srcFile, dst))
+            fileName = os.path.split(srcFile)[-1]
+            dstFile = os.path.join(dst, fileName)
+            if not os.path.exists(dstFile):
+                shutil.copy(srcFile, dst)
+                print("copying file: {0} -> {1}".format(srcFile, dst))
+            else:
+                print(
+                    "abandon copying! because file {0} has already been copied to {1}".
+                    format(srcFile, dst))
 
 
 if __name__ == "__main__":
